chore: refactored ui
